Velum Bootstrap: Don't wait for UI loaded
In cases where there are a large number of nodes, the page may take longer to load
and render than it takes for the first / single master to bootstrap. This is rare,
but I've seen it happen a few times.

So - Instead of waiting for all nodes to be "in progress", we just allow the check
below to wait for all nodes to be completed.