Make remote libvirt hosts scenario easier
I do all the development on my laptop but I cannot run a caasp cluster
because on it because it's not powerful enough.

However I do have a powerful workstation that I can use. By default
the devenv creates a cluster located inside of a natted network that
is reachable only by the virtualization host.

I could workaround that by using a ssh tunnel, but this is suboptimal,
especially when I want to play with applications running on top of the
kubernetes cluster.

With this change the VMs are still created inside of a dedicated network
managed by libvirt, but they are also reachable by the outside (once a
dedicated routing rule is added to the host).

For example, given the hypervisor host ip address is `192.168.1.39` and
all the machines are on the `192.168.200.0/24` network, the following
command will allow another host to access these VMs:

```
ip -4 route add 192.168.200.0/24 via 192.168.1.39
```

Changing the network type from `nat` to `route` is not enough: the
default subnet is used also by SUSE's R&D network. My laptop is always
connected to this network, hence all the requests to the
`10.17.0.0/22` network are automatically routed to the VPN instead of
my workstation.

This commit allows also to change the default subnet to use a different one that
is not conflicting with SUSE's R&D network (like the `172.17.0.0/22`
one).

Signed-off-by: Flavio Castelli <fcastelli@suse.com>