Make it possible to disable meltdown and spectre patches
This commit introduces a new flag that boots all the VMs with the
spectre and meltdown fixes disabled.

To do that we have to instruct terraform to boot the VMs with a series
of kernel params. We have also to provide libvirt a kernel and an initrd
files.

These files, plus the already present kernel boot parameters, are
extracted from the qcow image by a new helper tool.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>