Apply local changes to the haproxy config using cloud-init
Otherwise, local modifications done to the configuration of haproxy
never applies when we create a development environment. It's kind of
a corner case, but can be confusing if you are messing with the
`haproxy` configuration and creating development environments.

(cherry picked from commit cc6dfbed52102faf73e53a877e9768404d44d28e)