download-image: Find SHA256 by regex
Looking for SHA256 sums in files by specific columns fails if the files'
columns are not laid out to the get_sha256 function's assumptions.
Instead, look for a length-64 hex word in the file text by regex to find
the sha256 sum. Error if no sum is found or if more than one word
matches the regex.

Signed-off-by: Blaine Gardner <blaine.gardner@suse.com>