Revert "[testing] set js_errors to false"
Setting js_errors errors to false hides many real errors, and leads
to confusing results as rspec will just hang forever, eventually
timing out, instead of stopping and showing the failure.

The specific error mention in the original PR is a real bug in the
product, and it generally happens only very occasially. We shouldn't
just hide it!

This reverts commit 7443fe9b50e9cf09b903f2072727fcc1e0453749.