Add caasp-kvm release image to volume delete list
Leaving the "suse-caasp-whatever" KVM image in the volume list seems to "sometimes" prevent Terraform from assembling the cluster.  So, remove those images as well.
Also cleaned up the grep regex using a few `-e` options to make it a bit easier to read and extend in the future, and adjusted the xargs call for readability (and performance!). :)